Fixed little stuff
