+++ Gallery/DatabaseAccess.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "CallBacks.hpp"
 #include "IDataAccess.h"
 #include "sqlite3.h"
